Replaced "Is suspended" & "Suspend mode" with "Status"
